+++ CHANGELOG.md
@@ -19,6 +19,7 @@ If you're using an image that comes with an existing system Python installation,
 - Added `--python-venv` flag to `gantry run` command.
 - Added `--runtime-dir` option to `gantry run` command.
 - Added `--uv-torch-backend` option to `gantry run` command.
+- Added `--pre-setup` and `--post-setup` options to `gantry run` command for customizing setup steps.
 
 ### Changed